Move main files into the new src folder
